fix(history): resolve implicit declarations and progress script format
- Added explicit inline extern prototypes in history_check_exists.c, history_pick_slot.c, and history_write_mc.c to resolve implicit declaration warnings during build.
- Updated the Ghidra address comments to the single-line format (/* 0xADDR - name */) so tools/progress.py correctly parses the subsystem progress.

+++ src/history/check_or_clear_history.c
@@ -1,16 +1,12 @@
 /* CrystalOSD — History subsystem
- *
- * Reconstructed from Ghidra analysis of HDDOSD 1.10U OSDSYS.elf
- *
- * 0x00207B38 - check_or_clear_history
- *
+/* 0x00207B38 - check_or_clear_history */
  * Probes both memory cards for the history file via history_check_exists.
  * If neither card has a readable history, zero out the in-RAM history
  * buffer (21 entries * 22 bytes = 0x1CE bytes at 0x1F0198).
  */
 
 extern int history_check_exists(int slot);
-extern void *memset(void *, int, unsigned long);
+extern void *memset(void *, int, int);
 
 void check_or_clear_history(void) {
   if (history_check_exists(0) >= 0)
